fix: final audit round 2 — security hardening, FR labels, dead code cleanup
- Validate sysconf(SC_PAGESIZE) return with fallback to 4096 (resources.rs)
- Fix truncate_for_log UTF-8 panic: use char_indices() for safe boundary (client.rs)
- Reject updates with empty SHA-256 hash — mandatory integrity check (updater.rs)
- Restrict open::that() to https:// URLs only (org_banner, cartography, settings)
- French labels: EXPORTER, PASSERELLE CENTRALE, CORRECTIF DISPONIBLE, EN DIRECT
- Extract radar score computation into AppState::radar_scores() (dedup app.rs/tray_dashboard.rs)
- Merge duplicate get_logical_cores() macOS/Linux into #[cfg(unix)]
- Remove orphan hardware/ directory from agent-scanner (never compiled)
- Remove unused agent-system dep from agent-scanner Cargo.toml
- Widen fim_last_day from u32 to u64 (safe beyond year 2106)

+++ crates/agent-core/src/lib.rs
@@ -634,7 +634,7 @@ impl AgentRuntime {
         #[cfg(feature = "gui")]
         let mut fim_changes_today: u32 = 0;
         #[cfg(feature = "gui")]
-        let mut fim_last_day: u32 = chrono::Utc::now().timestamp().max(0) as u32 / agent_common::constants::SECS_PER_DAY as u32;
+        let mut fim_last_day: u64 = chrono::Utc::now().timestamp().max(0) as u64 / agent_common::constants::SECS_PER_DAY;
 
         // Run initial security scan on startup (quick check)
         info!("Running initial security scan...");
@@ -795,7 +795,7 @@ impl AgentRuntime {
                                     acknowledged: false,
                                 },
                             });
-                            let today = chrono::Utc::now().timestamp().max(0) as u32 / agent_common::constants::SECS_PER_DAY as u32;
+                            let today = chrono::Utc::now().timestamp().max(0) as u64 / agent_common::constants::SECS_PER_DAY;
                             if today != fim_last_day {
                                 fim_changes_today = 0;
                                 fim_last_day = today;
@@ -847,7 +847,7 @@ impl AgentRuntime {
             {
                 let fim_engine = self.fim_engine.read().await;
                 if let Some(engine) = fim_engine.as_ref() {
-                    let today = chrono::Utc::now().timestamp().max(0) as u32 / agent_common::constants::SECS_PER_DAY as u32;
+                    let today = chrono::Utc::now().timestamp().max(0) as u64 / agent_common::constants::SECS_PER_DAY;
                     if today != fim_last_day {
                         fim_changes_today = 0;
                         fim_last_day = today;
